% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\docType{data}
\name{constant_panelSubgroupMethods}
\alias{constant_panelSubgroupMethods}
\title{An internal constant thats repeated
user never need to call this}
\format{
An object of class \code{character} of length 2.
}
\usage{
constant_panelSubgroupMethods
}
\description{
An internal constant thats repeated
user never need to call this
}
\keyword{internal}
