% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\docType{data}
\name{default_generationTypes}
\alias{default_generationTypes}
\title{An internal constant thats repeated
user never need to call this}
\format{
An object of class \code{logical} of length 2.
}
\usage{
default_generationTypes
}
\description{
An internal constant thats repeated
user never need to call this
}
\keyword{internal}
