% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\docType{data}
\name{default_seedParams}
\alias{default_seedParams}
\title{An internal constant thats repeated
user never need to call this}
\format{
An object of class \code{list} of length 2.
}
\usage{
default_seedParams
}
\description{
An internal constant thats repeated
user never need to call this
}
\keyword{internal}
