% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{getDefaultTrainControl}
\alias{getDefaultTrainControl}
\title{getDefaultTrainControl}
\usage{
getDefaultTrainControl(m)
}
\arguments{
\item{m}{The method that the train control will be built for.}
}
\value{
a default caret trainControl.
}
\description{
Will return the default train control used for kfold or rsbmr.
}
