% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelValidation.R
\name{generateModelAndInfo}
\alias{generateModelAndInfo}
\title{Generate validation model information}
\usage{
generateModelAndInfo(i, modelParameters, method, getModel = F)
}
\arguments{
\item{i}{An integer to offset the seed by.}

\item{modelParameters}{A named list that has all arguments required to validate a panel.}

\item{method}{A string that identify what validation method to use.}

\item{getModel}{A logical to determine if you want a list of models for external use.}
}
\value{
A character vector of the model information if getModel is FALSE,
otherwise returns a list containing the model information in addition to the actual models.
}
\description{
The key function in validating a panel. It was designed so it can be ran in parallel,
so i acts as an seed offset when the function is called in parLapply.
}
