% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo.R
\name{generateModelParametersFromTwoBDPClass}
\alias{generateModelParametersFromTwoBDPClass}
\title{Generate model parameters to run key functions}
\usage{
generateModelParametersFromTwoBDPClass(bdp, method)
}
\arguments{
\item{bdp}{The TwoBDP object that will be parsed to generate a list of parameters to run a section of the TwoBDP algorithm.}

\item{method}{A String to identify what method the function will run through and compile the required arguments.}
}
\value{
A named list of required arguments needed to run a specific method.
}
\description{
Extract key arguments from TwoBDP object in order to run specific functions.
}
