% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelGeneration.R
\name{generatePanel}
\alias{generatePanel}
\title{Generate a panel for TwoBDP}
\usage{
generatePanel(i, modelParameters, maxFeatures, getModel = F)
}
\arguments{
\item{i}{An integer to offset seed value in modelParameters.}

\item{modelParameters}{A list of parameters that are used to generate a panel.}

\item{maxFeatures}{The max size of features required. Used as cutoff to select features from random forest.}

\item{getModel}{A logical to know if the model should be returned}
}
\value{
TwoBDP class with processed data appended.
}
\description{
Primary function that is used to generate a panel using random forest
}
