% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{get_mean_diff}
\alias{get_mean_diff}
\title{An internal function thats used to trim dataset
user never need to call this, can just call bdp$data after creating if needed}
\usage{
get_mean_diff(df, boolDex)
}
\description{
An internal function thats used to trim dataset
user never need to call this, can just call bdp$data after creating if needed
}
\keyword{internal}
