% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo.R
\name{getModelAndPred}
\alias{getModelAndPred}
\title{This function handles generating the validation models
Shouldnt be directly called, there are maco functions that handle this}
\usage{
getModelAndPred(seed, rngkind, fm, data, test, trainControl)
}
\description{
This function handles generating the validation models
Shouldnt be directly called, there are maco functions that handle this
}
\keyword{internal}
