% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo.R
\name{getModelInformation}
\alias{getModelInformation}
\title{This function handles getting the model information
Shouldnt be directly called, there are maco functions that use this}
\usage{
getModelInformation(fit, observedValues, numOfFeatures)
}
\description{
This function handles getting the model information
Shouldnt be directly called, there are maco functions that use this
}
\keyword{internal}
