% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo.R
\name{getValidationInformation}
\alias{getValidationInformation}
\title{This function handles getting the roc information
Shouldnt be directly called, there are maco functions that use this}
\usage{
getValidationInformation(roc)
}
\description{
This function handles getting the roc information
Shouldnt be directly called, there are maco functions that use this
}
\keyword{internal}
