% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_incorrectSize}
\alias{message_incorrectSize}
\title{An internal message function to allow clean code
user never need to call this}
\usage{
message_incorrectSize(param, sz, exsz, rnull = T)
}
\description{
An internal message function to allow clean code
user never need to call this
}
\keyword{internal}
