% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_missingValues_function}
\alias{message_missingValues_function}
\title{An internal message function to allow clean code
user never need to call this}
\usage{
message_missingValues_function(dataName, functionName, rnull = T)
}
\description{
An internal message function to allow clean code
user never need to call this
}
\keyword{internal}
