% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_missMatch_dataframe}
\alias{message_missMatch_dataframe}
\title{An internal message function to allow clean code
user never need to call this}
\usage{
message_missMatch_dataframe(comp, v1, v2, rnull = T)
}
\description{
An internal message function to allow clean code
user never need to call this
}
\keyword{internal}
