% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_NULL}
\alias{message_NULL}
\title{An internal message function to allow clean code
user never need to call this}
\usage{
message_NULL()
}
\description{
An internal message function to allow clean code
user never need to call this
}
\keyword{internal}
