% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{msg_c}
\alias{msg_c}
\title{An internal message function to allow clean code
user never need to call this}
\usage{
msg_c(v)
}
\description{
An internal message function to allow clean code
user never need to call this
}
\keyword{internal}
