% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{convertFlipList}
\alias{convertFlipList}
\title{convertFlipList}
\usage{
convertFlipList(lt)
}
\arguments{
\item{lt}{A list that will be converted into a dataframe.}
}
\value{
dataframe
}
\description{
this function is used to convert and transform a list of vectors into a dataframe.
}
