% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{createTempFileName}
\alias{createTempFileName}
\title{createTempFileName for file managment}
\usage{
createTempFileName(pre, num)
}
\arguments{
\item{pre}{A string that will be used as the prefix for the file.}

\item{num}{an integer that will be used to identify the data set.}
}
\value{
String file name
}
\description{
this function returns a temp file name that pads 0 before the number so the file output is uniformed and easy to manage.
}
