% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getIndex}
\alias{getIndex}
\title{getIndex managing index creation for train & test sets}
\usage{
getIndex(vec, ratio = 0.7, seed = 100, rngKind = "L'Ecuyer-CMRG", count = 1)
}
\arguments{
\item{vec}{A vector of values used to determine how the data will be split.}

\item{ratio}{A numeric value to know what indexes will be used in training data set.}

\item{seed}{A numeric value to set set the seed for reproducibility.}

\item{rngKind}{A string value to set set the RNGkind for reproducibility.}

\item{count}{An integer used to know how many indexes should be generated.}
}
\value{
A named list of indexes to subset a dataset into a train & test set.
}
\description{
this function is called at key parts in functions to manage reproducibility within functions.
Note that if count!=1 then random seeds will be generated using that number,
which are then used to split the data.
}
