% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{setRNG}
\alias{setRNG}
\title{setRNG managing reproducibility}
\usage{
setRNG(seed = 100, rngKind = "L'Ecuyer-CMRG")
}
\arguments{
\item{seed}{used to set seed in set.seed function.}

\item{rngKind}{used to set kind in RNGkind function.}
}
\description{
this function is called at key parts in functions to manage reproducibility within functions.
}
