% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{TwoBDP_addPanels}
\alias{TwoBDP_addPanels}
\alias{TwoBDP_addFeatureCombinations}
\title{TwoBDP_add functions}
\usage{
TwoBDP_addPanels(bdp, v)

TwoBDP_addFeatureCombinations(bdp, v)
}
\arguments{
\item{bdp}{A TwoBDP class that you want to manually add information to.}

\item{v}{A dataframe or vector depending if its used to skip the random forest or feature combinations}
}
\value{
TwoBDP object with the data append.
}
\description{
These functions were designed to allow the user to add their own dataframe of panels to skip random forest,
or their own vector of feature combinations to only perform validation.
}
\examples{
bdp=createTwoBDPObject(...)
bdp=TwoBDP_addPanels(bdp,v=df)
#or
features=c("feat1","feat1+feat2",...)
bdp=TwoBDP_addFeatureCombinations(bdp,v=features)

}
