% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{TwoBDP_makeFrequencyParameterList}
\alias{TwoBDP_makeFrequencyParameterList}
\alias{TwoBDP_makeValidationParameterList}
\title{TwoBDP_makeFrequencyParameterList}
\usage{
TwoBDP_makeFrequencyParameterList(
  amountOfFeatures = 1:10,
  append = NULL,
  sub = NULL,
  subGroupMethod = "all"
)

TwoBDP_makeValidationParameterList(
  cutoffs = c(AUC = 0.8, pvalue = 0.05),
  metric = "Accuracy",
  generationType = default_generationTypes,
  trainSplitRatio = 0.7,
  kfold = 10,
  repeats = c(kfold = 0, rsbmr = 10),
  validationMethods = c("kfold", "rsbmr"),
  tuneGrid = NULL,
  trainControls = NULL
)
}
\arguments{
\item{amountOfFeatures}{An integers vector that are the size of the combinations you want.}

\item{append}{A character vector of features you want appended to every sub panel of features identified.}

\item{sub}{A character vector of features you want to use to replace the tail of panels from. Panels that are equal or smaller than sub's length will be ignored.}

\item{subGroupMethod}{A string that is either "all" or "slidingWindow" to determine how feature combinations are performed.}

\item{cutoffs}{A named numeric vector containing the "AUC" cutoff and "pvalue" cutoff used to trim results by.}

\item{metric}{A string for the metric used in caret Train function.}

\item{generationType}{A named logical vector that determines if "data" and "model" seeds are statically or dynamically used.}

\item{trainSplitRatio}{A numeric value used to split the data to a training data set. The remaining values will be used to validate.}

\item{kfold}{A numeric value used for how many folds "kfold" method will perform.}

\item{repeats}{A numeric named vector that determines how many repeats are performed for each validation method. Only effects "rsbmr".}

\item{validationMethods}{A string vector used to identify all validation methods to run. Currently "kfold", and "rsbmr" are the only options.}

\item{tuneGrid}{A caret valid tuneGrid used for all validation methods.}

\item{trainControls}{A list of caret trainControls used for each validation method.}
}
\value{
List of named parameters that will be called in frequency mapping and feature combination.

List of named parameters that will be called in frequency mapping and feature combination.
}
\description{
TwoBDP_makeFrequencyParameterList

TwoBDP_makeValidationParameterList
}
