% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{TwoBDP_makePanelParameterList}
\alias{TwoBDP_makePanelParameterList}
\title{TwoBDP_makePanelParameterList}
\usage{
TwoBDP_makePanelParameterList(
  totalPanels = 2000,
  topPanels = 200,
  generationType = default_generationTypes,
  rfInputDataSize = 1,
  maxPanelSize = 10,
  trainSplitRatio = 1,
  ntree = 500,
  nodeSize = 1,
  metric = "Accuracy",
  tuneGrid = NULL,
  trainControl = NULL
)
}
\arguments{
\item{totalPanels}{Data frame of features "columns" and samples "rows" to analyze. Rownames should be sample identifiers which are all in the rownames in metadata.}

\item{topPanels}{Data frame which has a 2 factor column which will be used as the point of interest.}

\item{generationType}{A string path where results will be saved.}

\item{rfInputDataSize}{A numeric value to determine if the data should be subset before running 2bdp.}

\item{maxPanelSize}{A string for the column name in metadata, which is the 2 factor for analysis.}

\item{trainSplitRatio}{A string for the non-control value in decisionPointColumn.}

\item{ntree}{A named list that contains the "seed" and "RNGkind".}

\item{nodeSize}{An int for the amount of cores use in parallel.}

\item{metric}{A list of required parameters to run the random forest. See TwoBDP_makePanelParameterList function for more details.}

\item{tuneGrid}{A list of required parameters to run the frequency map. See TwoBDP_makeFrequencyParameterList function for more details.}

\item{trainControl}{A list of required parameters to run the validation. See TwoBDP_makeValidationParameterList function for more details.}
}
\value{
List of named parameters that will be called in panelGeneration function.
}
\description{
TwoBDP_makePanelParameterList
}
