% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyMap.R, R/panelGeneration.R,
%   R/panelValidation.R
\name{frequencyMap}
\alias{frequencyMap}
\alias{PanelExpansion}
\alias{panelGeneration}
\alias{panelValidation}
\title{Two BDP Automation}
\usage{
frequencyMap(bdp)

PanelExpansion(bdp)

panelGeneration(batchSize = 5, bdp)

panelValidation(batchSize = 10, bdp)
}
\arguments{
\item{bdp}{The TwoBDP object to process}

\item{batchSize}{How many datasets to process before saving. Only meaningful when progress is saved.}
}
\value{
TwoBDP class with processed data appended.
}
\description{
These functions are used to automate, parallize, and process the data in batches.
Once the function has been completed, the TwoBDP object will then have the required data
needed for the next step.
Order of functions are as follows:
bdp=createTwoBDPObject(bdp)
bdp=panelGeneration(bdp)
bdp=frequencyMap(bdp);bdp=PanelExpansion(bdp)
bdp=panelValidation(bdp)
}
\examples{
df=read.csv("../TwoBDP/TestData/df.csv")
md=read.csv("../TwoBDP/TestData/md.csv")
rownames(df)=df[,1];df=df[,-1]
colnames(df)=paste("V",1:ncol(df),sep="")
rownames(md)=md[,1]

bdp=createTwoBDPObject(data = df,metadata = md,
  decisionPointColumn="poi",pointOfInterest="TRUE",threads = 10,saveDirectory = ".",
  panelGenerationParameters=TwoBDP_makePanelParameterList(totalPanels=100)

bdp=panelGeneration(batchSize = 50,bdp = bdp)
bdp=frequencyMap(bdp = bdp)
bdp=PanelExpansion(bdp = bdp)
bdp=panelValidation(bdp = bdp)

}
