% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoBDPClass.R
\name{createTwoBDPObject}
\alias{createTwoBDPObject}
\title{createTwoBDPObject}
\usage{
createTwoBDPObject(
  data = NULL,
  metadata = NULL,
  saveDirectory = NULL,
  decisionPointColumn = NULL,
  pointOfInterest = NULL,
  seedParms = default_seedParams,
  threads = 1,
  panelGenerationParameters = TwoBDP_makePanelParameterList(),
  frequencyMapParameters = TwoBDP_makeFrequencyParameterList(),
  validationParameters = TwoBDP_makeValidationParameterList()
)
}
\arguments{
\item{data}{Data frame of features "columns" and samples "rows" to analyze. Rownames should be sample identifiers which are all in the rownames in metadata.}

\item{metadata}{Data frame which has a 2 factor column which will be used as the point of interest.}

\item{saveDirectory}{A string path where results will be saved.}

\item{decisionPointColumn}{A string for the column name in metadata, which is the 2 factor for analysis.}

\item{pointOfInterest}{A string for the non-control value in decisionPointColumn.}

\item{seedParms}{A named list that contains the "seed" and "RNGkind". see ?RNGkind for more details and options.}

\item{threads}{An int for the amount of cores use in parallel.}

\item{panelGenerationParameters}{A list of required parameters to run the random forest. See TwoBDP_makePanelParameterList function for more details.}

\item{frequencyMapParameters}{A list of required parameters to run the frequency map. See TwoBDP_makeFrequencyParameterList function for more details.}

\item{validationParameters}{A list of required parameters to run the validation. See TwoBDP_makeValidationParameterList function for more details.}
}
\value{
TwoBDP class that will be uses for all TwoBDP related functions.
}
\description{
createTwoBDPObject
}
\examples{
bdp=createTwoBDPObject(data=df,metadata=metaDF,
decisionPointColumn="metadataColumn",pointOfInterest="Case")

}
